Fix for dropped txs in GN's tx broadcast (#471)
* Fix for dropped txs in GN's tx broadcast

This patch fixes the bug that causes some txs to remain only in the RN's
pending pool when they are dropped in peer.AsyncSendTransactions(), which is
called in ProtocalManager.BroadcastTxs.  The fix is to use peer.SendTransactions(),
which is a synchronous call, instead of AsyncSendTransaction for non-CN node
types.  Note that this fix is not a 100% solution, because synchronous sends
may fail but this fix does not handle that case.

For issue #465

Signed-off-by: Sangmin Seo <sam.seo@groundx.xyz>

* Print error message when peer.SendTransactions fails