Introduce SerializeForSignToBytes() to optimize tx signing (#1523)
SerializeForSign() is used to generate a hash which is also used for the tx
signature.  As we previously found that RLP serialization of []interface is
very slow, this commit introduces a function SerializeForSignToBytes for
all new transaction types. SerializeForSignToBytes returns a slice of
RLP-encoded bytes instead of returning []interface{}.

This improves performance of tx.Hash() by 15.7% on average.

Closes #1433

Signed-off-by: Junghyun Colin Kim <colin.kim@groundx.xyz>