Fixed simple table test (#2857)
- The Lookup test is for unit test in
order to check if the function works supposedly. However, whenever new
simple table is created, it creates a loop and it might be possible to
change the contents of the table which is not for this unit test.
- It fixes the failure case of the test by calling the function close
right after the simple table instance is created so that the Lookup
function is only being checked.