[homi] Introduce option num-validators (#3278)
We have needs that the number of total CNs is different from
the number of total validators. To meet the needs, this commit
introduces num-validators.

If you run `homi --num 7 --num-validators 4`,
seven nodekeys and validators will be generated, but
only first four nodekeys will be included in the validator list.